Path optimization and fixing of typo in station names
I have optimized the search algorithm by changing it from a BFS to a DFS and
changing it so that the stations on the same line will be put on top of the stack
after everone else (and thus will pop first).

Currently, it's working very well. I will have to do some massive cleaning up of
find_metro_path.py however, as it is currently a mess and it uses "queue" in its
variable name instead of "stack", and that would just cause confusion.

 Changes to be committed:
	modified:   .gitignore
	modified:   build_metro.py
	modified:   find_metro_path.py
	modified:   metro_lines.py